Move CWS phase management functions to dedicated module
