+++ cms/server/contest/handlers/taskusertest.py
@@ -52,11 +52,13 @@ from cms.db import Task, UserTest, UserTestFile, UserTestManager, \
     UserTestResult
 from cms.grading.languagemanager import get_language
 from cms.grading.tasktypes import get_task_type
-from cms.server import actual_phase_required, multi_contest
+from cms.server import multi_contest
 from cmscommon.archive import Archive
 from cmscommon.crypto import encrypt_number
 from cmscommon.datetime import make_timestamp
 from cmscommon.mimetypes import get_type_for_file_name
+
+from ..phase_management import actual_phase_required
 
 from .contest import ContestHandler, FileHandler, NOTIFICATION_ERROR, \
     NOTIFICATION_SUCCESS
